requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2==2.9.5
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
WTForms==3.0.1
email-validator==1.3.1
